+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [Unreleased]
 
 - Add `base-address-shift` config flag
+- Use `PascalCase` for type idents, fix case changing bugs, add `--ident-format` (`-f`) option flag
 
 ## [v0.31.5] - 2024-01-04